Hw04 auth

+++ services/contactsService.js
@@ -1,29 +1,32 @@
-const { Contacts } = require("../db");
+const { Contacts } = require("../models");
 
-const listContacts = async () => {
-  return await Contacts.find({});
+const listContacts = async (owner, skip, limit, favorite) => {
+  return await Contacts.find(favorite ? { owner, favorite } : { owner })
+    .populate("owner", "email")
+    .skip(skip)
+    .limit(limit);
 };
 
-const getContactById = async (contactId) => {
-  return await Contacts.findById({ _id: contactId });
+const getContactById = async (contactId, owner) => {
+  return await Contacts.findById({ _id: contactId, owner });
 };
 
 const addContact = async (body) => {
   return await Contacts.create(body);
 };
 
-const removeContact = async (contactId) => {
-  return await Contacts.findByIdAndDelete({ _id: contactId });
+const removeContact = async (contactId, owner) => {
+  return await Contacts.findByIdAndDelete({ _id: contactId, owner });
 };
 
-const updateContact = async (contactId, body) => {
-  return await Contacts.findByIdAndUpdate({ _id: contactId }, body, {
+const updateContact = async (contactId, body, owner) => {
+  return await Contacts.findByIdAndUpdate({ _id: contactId, owner }, body, {
     new: true,
   });
 };
 
-const updateStatusContact = async (contactId, body) => {
-  return await Contacts.findByIdAndUpdate({ _id: contactId }, body, {
+const updateStatusContact = async (contactId, body, owner) => {
+  return await Contacts.findByIdAndUpdate({ _id: contactId, owner }, body, {
     new: true,
   });
 };
